+++ packages/api/src/routers/index.ts
@@ -1,5 +1,6 @@
 import { protectedProcedure, publicProcedure, router } from "../index";
 import { channelsRouter } from "./channels";
+import { jobsRouter } from "./jobs";
 import { plexRouter } from "./plex";
 import { sourcesRouter } from "./sources";
 import { usersRouter } from "./users";
@@ -15,6 +16,7 @@ export const appRouter = router({
     };
   }),
   channels: channelsRouter,
+  jobs: jobsRouter,
   plex: plexRouter,
   sources: sourcesRouter,
   users: usersRouter,